bug fix in graph search engine gui : the search for nodes and edges according to a metric was erroneous due to a bad conversion of the reference value (integer instead of double)
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@10545 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ plugins/perspective/GraphPerspective/src/SearchWidget.cpp
@@ -284,8 +284,8 @@ void SearchWidget::search() {
 
     if (isNumericComparison()) {
       DoubleProperty* doubleProp = new DoubleProperty(g);
-      doubleProp->setAllNodeValue(_ui->tableWidget->item(0, 0)->data(Qt::DisplayRole).toInt());
-      doubleProp->setAllEdgeValue(_ui->tableWidget->item(0, 0)->data(Qt::DisplayRole).toInt());
+      doubleProp->setAllNodeValue(_ui->tableWidget->item(0, 0)->data(Qt::DisplayRole).toDouble());
+      doubleProp->setAllEdgeValue(_ui->tableWidget->item(0, 0)->data(Qt::DisplayRole).toDouble());
       b = doubleProp;
     }
     else {
